refactor(cost): organize timeline application
