+++ src/services/portfolio_transaction_processing_service/app/application/cost_basis_processing/__init__.py
@@ -10,17 +10,20 @@ from .preparation import (
     PreparedCostTransaction,
     prepare_cost_transaction,
 )
+from .timeline import CostBasisTimelineProcessor, build_cost_basis_timeline_processor
 from .upstream_cash_leg import UpstreamCashLegUnavailableError, validate_upstream_cash_leg
 
 __all__ = [
     "AverageCostPoolRebuildPlanner",
     "CostBasisCalculationResult",
+    "CostBasisTimelineProcessor",
     "CostProcessingRoute",
     "FxRateNotFoundError",
     "InstrumentReferenceUnavailableError",
     "OpenLotPersistenceScope",
     "PreparedCostTransaction",
     "UpstreamCashLegUnavailableError",
+    "build_cost_basis_timeline_processor",
     "enrich_cost_basis_transactions_with_fx",
     "prepare_cost_transaction",
     "persist_open_lot_state",
